preparing to improve delete functionality for registration. Multiple delete will be supported
