+++ src/main/js/my-app/src/utils/isJsonString.js
@@ -1,0 +1,8 @@
+export default str => {
+  try {
+    JSON.parse(str);
+  } catch (e) {
+    return false;
+  }
+  return true;
+};
